requirements.txt
fastapi
uvicorn
pydantic
requests
faiss-cpu
python-dotenv
langchain
langchain-google-genai
PyMuPDF
pdfplumber
pytesseract
